Create ZS_StoneMasonKarel
Instructions
Karel has been hired to build the columns in the Temple of Artemis in Efes. In particular, there are a set of arches where the stones (represented by beepers, of course) are missing from the columns supporting the arches, as follows:
When Karel is done, the missing columns should be replaced by beepers, so the final picture would look like this:

Karel may count on the following facts about the world, listed below:

Karel starts at bottom left corner, facing right (aka east).

The columns are exactly four squares apart, on the 1st, 5th, 9th, and 13th columns.

Karel can assume that columns are always five units high.

Your program will be much easier to write and easier to read, if you use for loops. There are multiple opportunities to use for loops in this problem!

In addition to for loops, you should be practicing defining functions. For example, it would make sense to have a build_column function.