RandomNumbers
The starter code for this program prints out a single random number in the range 1 to 100, inclusive.

Modify this program so that instead of printing one random number, it prints 10. To repeat a block of code a 10  times, you should use a for loop:

for i in range(10):

Each run of your program will be different, because your program is using a pseudo random number generator. Here is one example run:

54

65

78

63

80

23

92

53

23

40

Unlike other programs, this one does not have automatic tests. When you determine that you have completed the problem hit the blue button on the right which says: "Mark Complete"